Bot trained on predicting blacks win probability
Status: Working
A real evaluation is still missing as we first need to implement
end-of-game stuff again
Trained on using the current board as input and the outcome of the game
as output, where it is 1 if black wins and 0 if black loses.
Plays by comparing the current win probability of black with the
probabilities after each possible move. Choses the best move, and passes
if no move can make the current situation better (raise the probability
if player == black or lower it if player == white)

The cool thing is that this kind of fixes `pass` prediction, with which
the NN bot struggled.