Process SQL-Query results at once instead by row
Changes on both BaseLearn and Learn.py
Instead of `handle_row` I propose `handle_data`, where we handle
directly the complete data that we get from the database.
To be performant I strongly encourage you to only use native numpy
functions, but this still leaves the option to iterate over the rows of
the data in case that's needed.

This serves as a proposal to change our current data-handling. I did not
recreate the complete functionality that was in the Learn.py (sorry
Ben!) and I do not know how to implement symmetries in a vectorized form
without looping.