Error types now have their own file
Defining an error class in `Game.py` and importing that to `Move.py`
created some weird behaviour of circular imports. `Move.py` needs to be
able to raise Exceptions, as we do not want the program to crash for bad
user inputs.

Maybe a different location would be more suitable.