`bots` update
Changes to all of them, but probably the major change was to use
pandas.DataFrame to get back the results of the sql query, instead of
loading it into a numpy array by default, as numpy arrays only accept a
single data-type for all entries at once. As the `result` is saved as a
string the whole array contained therefore strings, which is highly
inefficient seing as the rest of it are only integers.