Updates to learn.py
Now uses the `model_zero.py` network, inspired by AlphaGO-Zero.
Learns both a value and a policy prediction at the same time,
by using two heads and propagating back the sum of the errors.

Also updated the loop to be visually more appealing, as well
as to track more statistics that could be useful for the evaluation
after the training process.